Move internal/verdicts.py -> internal/verify/verdicts_parser.py
